pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "perteval"
version = "0.1.0"
description = "A unified evaluation framework for single-cell perturbation prediction"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "anndata>=0.10",
    "numpy>=1.24",
    "polars>=0.20",
    "pyyaml>=6.0",
    "scipy>=1.11",
]

[project.optional-dependencies]
models = ["torch>=2.0"]
data = ["datasets>=2.0"]
all = ["perteval[models,data]"]
dev = [
    "pytest>=8.0",
    "ruff>=0.11",
]

[project.scripts]
perteval = "perteval.cli.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["N803", "N806"]  # Allow uppercase X, Y, XX etc. — standard math/matrix conventions

[tool.pytest.ini_options]
testpaths = ["tests"]
